requirements.txt
flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
PyJWT
cryptography
argon2-cffi
gunicorn
python-dotenv
flask-restx
pymysql